requirements.txt
h5py==3.1.0
numpy==1.19.0
opencv-python==4.4.0.46
Pillow==8.0.1
scipy==1.5.4
matplotlib==3.3.3
